+++ apps/content-analysis/src/utils/getMorphologyData.js
@@ -6,26 +6,30 @@ let morphologyData = null;
  * @returns {Object} The morphology data, or an empty object if not available.
  */
 function loadLocalMorphologyData() {
-	let data, dataDe, dataNL, dataES, dataFR, dataPL = {};
+	let data, dataDe, dataNL, dataES, dataFR, dataRU, dataIT, dataPL = {};
 	try {
 		// Disabling global require to be able to fail.
 		// eslint-disable-next-line global-require
 		data = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-v3.json" );
 		// eslint-disable-next-line global-require
-		dataDe = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-de-v5.json" );
+		dataDe = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-de-v9.json" );
 		// eslint-disable-next-line global-require
-		dataNL = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-nl-v5.json" );
+		dataNL = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-nl-v9.json" );
 		// eslint-disable-next-line global-require
-		dataES = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-es-v5.json" );
+		dataES = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-es-v9.json" );
 		// eslint-disable-next-line global-require
-		dataFR = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-fr-v5.json" );
+		dataFR = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-fr-v9.json" );
 		// eslint-disable-next-line global-require
-		dataPL = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-pl-v6.json" );
+		dataRU = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-ru-v9.json" );
+		// eslint-disable-next-line global-require
+		dataIT = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-it-v9.json" );
+		// eslint-disable-next-line global-require
+		dataPL = require( "../../../../packages/yoastseo/premium-configuration/data/morphologyData-pl-v9.json
 	} catch ( error ) {
 		// Falling back to empty data.
 	}
 
-	return merge( data, dataDe, dataNL, dataES, dataFR, dataPL );
+	return merge( data, dataDe, dataNL, dataES, dataFR, dataRU, dataIT, dataPL );
 }
 
 /**