+++ packages/yoastseo/src/morphology/morphoHelpers/regexHelpers.js
@@ -25,3 +25,18 @@ export function searchAndReplaceWithRegex( word, groupOfRegexAndReplacements ) {
 		}
 	}
 }
+
+/**
+ * Loops through a nested array of regexes and replacement pairs, and applies all replacements for which a match is found.
+ *
+ * @param {string}      word                        The word that may need to be modified.
+ * @param {string[]}    listOfRegexAndReplacement   The array with the regexes and the replacements.
+ *
+ * @returns {string} The modified word.
+ */
+export function applyAllReplacements( word, listOfRegexAndReplacement ) {
+	listOfRegexAndReplacement.forEach( function( setOfRegexAndReplacement ) {
+		word = word.replace( new RegExp( setOfRegexAndReplacement[ 0 ] ), setOfRegexAndReplacement[ 1 ] );
+	} );
+	return word;
+}