+++ packages/yoastseo/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file. Releases without a changelog entry contain only minor changes that are irrelevant for users of this library.
 We will follow [Semantic Versioning](http://semver.org/) from version 2 and onwards.
 
+## 1.78.0 July 6th, 2020
+### Enhancements
+* Adds specs for irregular verbs for the full forms exception list.
+* Adds check for verbs on -ions before stemming the verb suffix -ons and does not stem -ons if the word ends on -ions (for these words, the suffix is most likely -s, not -ons).
+* Adds "кроме того" to Russian transition words.
+* Adds a check for whether the word belongs to stemming exceptions to the Russian stemmer.
+* Adds an exception check for irregular Italian diminutives to the Italian stemmer and removes _scrivere_, _scrive_, and _scritto_ from function words.
+* Adds a check for exception lists of full forms and words with multiple stems to the Russian stemmer.
+* Adds some irregular plural forms of Italian words to function words and morphologyData file, including specs and a function to call the exception list.
+* Changes the `getProminentWordsForInternalLinking` research so that it only runs for texts over 400 words or texts over 300 words with a title and/or a metadescription specified.
+* Changes the output of the research so that it not only returns the list of prominent words but also information about the presence of metadescription and title, as well as the length of text. This information is later used to return a customized message to the user, within the internal linking suggestions container.
+
+## 1.77.0 June 22nd, 2020
+### Enhancements
+* Adds check for stemming -ons suffix in French, and for the exception list of words where only -s should be stemmed in French.
+* Adds check for nouns in Russian on _ость_ before removing verb suffixes (including the newly added suffix _ть_).
+
+## 1.76.0 June 8th, 2020
+### Enhancements
+* Adds a check for the exception list of French verbs with multiple stems and stems them by returning the indicated canonical stem.
+* Adds a stemmer for the Italian language.
+* Adds an exception check for words ending in -is/us/os where -s should not be stemmed.
+* Improves the way keyphrases containing words ending in "ent" are recognized in the text.
+* Stems French words that are considered too short to be stemmed according to the stemming rules, but that should nevertheless be stemmed.
+
+## 1.75.0 May 25th, 2020
+### Enhancements
+* Adds a stemmer for the Russian language.
+* Adds checks for exception list with full forms and exception list of words with multiple stems to the French stemmer.
+* Checks an exception list of plurals with -x suffix and stems them correctly.
+* Transfers data from French stemmer to data file and improves stemming of words in -issement in French.
+
 ## 1.74.0 May 11th, 2020
 ### Added
 * Adds French to the list of languages for which we have morphology support.